refactor: for dialog best practice
